Feature: all: avoid Year 2038 issues
Pacemaker has generally serialized time_t values in XML using int, which
creates a Year 2038 problem. This commit uses long long where possible.

Additionally, the result of the time(3) function was not always stored as
time_t. Now, it is, with the exception of the (currently unsigned int) storage
of time values in lrmd_event_data_t, to retain API backward compatibility.
Those should be changed to time_t at the next API breakage. At least they're
unsigned, so they are a year 2106 problem. :-)

This does not ensure that all time differences are also stored as time_t, which
is likely unnecessary but would probably be a good idea.

It is possible there are other uses of epoch time in XML, or storage of epoch
time in something other than time_t, that I didn't catch. The code will need
more thorough investigation.

+++ daemons/controld/controld_remote_ra.c
@@ -356,10 +356,10 @@ report_remote_ra_result(remote_ra_cmd_t * cmd)
     op.interval_ms = cmd->interval_ms;
     op.rc = cmd->rc;
     op.op_status = cmd->op_status;
-    op.t_run = cmd->start_time;
-    op.t_rcchange = cmd->start_time;
+    op.t_run = (unsigned int) cmd->start_time;
+    op.t_rcchange = (unsigned int) cmd->start_time;
     if (cmd->reported_success && cmd->rc != PCMK_OCF_OK) {
-        op.t_rcchange = time(NULL);
+        op.t_rcchange = (unsigned int) time(NULL);
         /* This edge case will likely never ever occur, but if it does the
          * result is that a failure will not be processed correctly. This is only
          * remotely possible because we are able to detect a connection resource's tcp
@@ -505,7 +505,7 @@ synthesize_lrmd_success(lrm_state_t *lrm_state, const char *rsc_id, const char *
     op.op_type = op_type;
     op.rc = PCMK_OCF_OK;
     op.op_status = PCMK_LRM_OP_DONE;
-    op.t_run = time(NULL);
+    op.t_run = (unsigned int) time(NULL);
     op.t_rcchange = op.t_run;
     op.call_id = generate_callid();
     process_lrm_event(lrm_state, &op, NULL, NULL);
