+++ lib/pacemaker/pcmk_sched_transition.c
@@ -140,7 +140,7 @@ create_op(xmlNode *cib_resource, const char *task, guint interval_ms,
     op->rc = outcome;
     op->op_status = 0;
     op->params = NULL;          /* TODO: Fill me in */
-    op->t_run = time(NULL);
+    op->t_run = (unsigned int) time(NULL);
     op->t_rcchange = op->t_run;
 
     op->call_id = 0;
